change from ssize_t to size_t and long long to ensure usability on 32-bit systems

+++ src/timsort.cpp
@@ -4,17 +4,23 @@
 #include "timsort.hpp"
 
 namespace ts {
+
+    // All instances of static_cast<long long> are fine, as our data will never be larger than 500,000 elements.
+    // long long can represent at least 9,223,372,036,854,775,807.
     void binaryInsertionSort(std::vector<Game*>& games, bool (*comparator)(const Game* lhs, const Game* rhs)) {
-        for (ssize_t leftUnsorted = 1; leftUnsorted < static_cast<ssize_t>(games.size()); leftUnsorted++) {
+        for (size_t leftUnsorted = 1; leftUnsorted < games.size(); leftUnsorted++) {
             // The key is the first unsorted element
             Game* key = games[leftUnsorted];
 
             // Search for the appropriate key location with binary search using std::lower_bound for sort stability
-            const ssize_t targetIndex = (std::lower_bound(games.begin(), games.begin() + leftUnsorted, key, comparator) - games.begin());
+            const size_t targetIndex = (std::lower_bound(games.begin(),
+                                                         games.begin() + static_cast<long long>(leftUnsorted), key,
+                                                         comparator) - games.begin());
 
             // Shift elements to the right if necessary
             if (targetIndex < leftUnsorted) {
-                std::shift_right(games.begin() + targetIndex, games.begin() + leftUnsorted + 1, 1);
+                std::shift_right(games.begin() + static_cast<long long>(targetIndex), games.begin() +
+                                     static_cast<long long>(leftUnsorted) + 1, 1);
             }
 
             // Move the key to the correct position
@@ -54,34 +60,37 @@ namespace ts {
     // Technically an introspective merge/insertion sort, but it's essentially a simplified timsort (no gallop)
     void timsort(std::vector<Game*>& games, bool (*comparator)(const Game* lhs, const Game* rhs)) {
         // Minimum size of a run. 512 is an arbitrary value, but it seems to work well
-        // Cast to ssize_t when using iterators to resolve narrowing conversion warnings when adding offsets
-        // This is fine, since ssize_t is equal in size to long long int
-        constexpr ssize_t RUN_SIZE = 512;
+        // Cast to size_t when using iterators to resolve narrowing conversion warnings when adding offsets
+        // This is fine, since size_t is equal in size to long long int
+        constexpr size_t RUN_SIZE = 512;
 
-        const ssize_t gameSize = static_cast<ssize_t>(games.size()); // NOLINT(*-use-auto)
+        const size_t gameSize = static_cast<size_t>(games.size()); // NOLINT(*-use-auto)
         // Sort individual vector slices
-        for (ssize_t i = 0; i < gameSize; i += RUN_SIZE) {
-            std::vector slice(games.cbegin() + i, games.cbegin() + std::min(i + RUN_SIZE, gameSize));
+        for (size_t i = 0; i < gameSize; i += RUN_SIZE) {
+            std::vector slice(games.cbegin() + static_cast<long long>(i),
+                              games.cbegin() + static_cast<long long>(std::min(i + RUN_SIZE, gameSize)));
             binaryInsertionSort(slice, comparator);
             // Use copy to quickly move the result back into the original vector
-            std::ranges::copy(slice.begin(), slice.end(), games.begin() + i);
+            std::ranges::copy(slice.begin(), slice.end(), games.begin() + static_cast<long long>(i));
         }
 
         // Start merging the vectors
-        for (ssize_t size = RUN_SIZE; size < gameSize; size *= 2) {
-            for (ssize_t left = 0; left < gameSize; left += 2 * size) {
-                const ssize_t mid = std::min(left + size, gameSize);
-                const ssize_t right = std::min(left + 2 * size, gameSize);
+        for (size_t size = RUN_SIZE; size < gameSize; size *= 2) {
+            for (size_t left = 0; left < gameSize; left += 2 * size) {
+                const size_t mid = std::min(left + size, gameSize);
+                const size_t right = std::min(left + 2 * size, gameSize);
 
                 // Only start merging if the right slice is of non-zero size
                 if (mid < right) {
                     // Create the slices to merge
-                    std::vector leftSlice(games.cbegin() + left, games.cbegin() + mid);
-                    std::vector rightSlice(games.cbegin() + mid, games.cbegin() + right);
+                    std::vector leftSlice(games.cbegin() + static_cast<long long>(left),
+                                          games.cbegin() + static_cast<long long>(mid));
+                    std::vector rightSlice(games.cbegin() + static_cast<long long>(mid),
+                                           games.cbegin() + static_cast<long long>(right));
                     std::vector<Game*> result(right - left);
                     merge_(leftSlice, rightSlice, result, comparator);
                     // Use copy to quickly move the result back into the original vector
-                    std::ranges::copy(result.begin(), result.end(), games.begin() + left);
+                    std::ranges::copy(result.begin(), result.end(), games.begin() + static_cast<long long>(left));
                 }
             }
         }
